+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/MagicPlugin.java
@@ -32,6 +32,7 @@ import com.elmakers.mine.bukkit.magic.command.MagicCommandExecutor;
 import com.elmakers.mine.bukkit.magic.command.MagicConfigCommandExecutor;
 import com.elmakers.mine.bukkit.magic.command.MagicEditorCommandExecutor;
 import com.elmakers.mine.bukkit.magic.command.MagicGiveCommandExecutor;
+import com.elmakers.mine.bukkit.magic.command.MagicHelpCommandExecutor;
 import com.elmakers.mine.bukkit.magic.command.MagicItemCommandExecutor;
 import com.elmakers.mine.bukkit.magic.command.MagicKitCommandExecutor;
 import com.elmakers.mine.bukkit.magic.command.MagicMapCommandExecutor;
@@ -205,6 +206,7 @@ public class MagicPlugin extends JavaPlugin implements MagicAPI
         new MagicWarpCommandExecutor(controller).register(this);
         new MagicConfigCommandExecutor(this, controller).register(this);
         new MagicEditorCommandExecutor(this, controller).register(this);
+        new MagicHelpCommandExecutor(this).register(this);
         new ArenaCommandExecutor(this, controller.getArenas()).register(this);
         // Note: The Citizens trait command is registered later in MagicController if we decide it's needed
     }